refactor(billing): 拆分 3735 行单体页为组件并统一四路内联编辑（Phase D1）
- system/billing/page.tsx 3735 行 → 152 行装配层，渲染逻辑拆入 src/components/admin/billing/*（11 个文件）：billing-format 纯函数、use-billing-price-row-edit 统一编辑状态机、summary-cards / unresolved-models / price-catalog / logs-card / override-reset-dialog / price-row / tier-sub-table / mobile-tier-list / expand-chevron
- 四路价目编辑（synced / override / tier-only 桌面行 + 移动卡）统一消费 useBillingPriceRowEdit（editTarget + draft + save/cancel/validate），mutation 实例外部传入共享 pending 态；桌面阶梯子表与移动阶梯列表各自三处逐字重复合并为单组件
- IA 不动；morph-billing-override-reset 接线逐字保留（enter/exit 与 morphSourceRef 留在 page.tsx）
- 契约对齐：管理头卡改用共享 PageHeader/PageShell/IconBox；4 张指标卡数字补 tabular-nums；半径全程 cf token
- 有意偏差：移动端 override/synced 卡补齐原仅 tier-only 卡有的 Enter/Escape 键盘行为（handler 超集，HTML 不变）
- 测试：billing-format 与 use-billing-price-row-edit 单测 41 用例；admin-page-mocks 补 billing 全系端点 fixture；visual/a11y 各加 billing 页条目，billing 视觉基线首次生成（就绪锚点用仅渲染一次的 custom-unpriced-model，避开桌面/移动双渲染的 gpt-4.1）

验证：tsc/lint 零错误；radius 守护 7/7；新单测 48 绿；全量 test:run 3181 绿；e2e --workers=2 20 绿 + 1 已知 flaky（logs-routing-decision）单跑复验绿（billing-tier-flow 全程守护通过）；visual 7/7（既有基线零漂移）、a11y 14/14；双主题截图（全页/内联编辑/重置弹窗）经 Opus 评审 GO；reset morph 冒烟走通（创建 visual-smoke-d1 覆盖价 → reset 确认 → 自清理）。
